feat: Add shadcn/ui CLI integration, Move Alert component to components/ui
