--- packages/webapp-libs/webapp-core/src/components/alert/alert.component.tsx
@@ -1,13 +1,0 @@
-import { type VariantProps } from 'class-variance-authority';
-import * as React from 'react';
-
-import { cn } from '../../lib/utils';
-import { alertVariants } from './alert.styles';
-
-export const Alert = React.forwardRef<
-  HTMLDivElement,
-  React.HTMLAttributes<HTMLDivElement> & VariantProps<typeof alertVariants>
->(({ className, variant, ...props }, ref) => (
-  <div ref={ref} role="alert" className={cn(alertVariants({ variant }), className)} {...props} />
-));
-Alert.displayName = 'Alert';
