--- packages/webapp-libs/webapp-core/src/components/alert/alert.styles.ts
@@ -1,18 +1,0 @@
-import { cva } from 'class-variance-authority';
-
-export const alertVariants = cva(
-  'relative w-full rounded-lg border p-4 [&>svg]:absolute [&>svg]:text-foreground [&>svg]:left-4 [&>svg]:top-4 [&>svg+div]:translate-y-[-3px] [&:has(svg)]:pl-11',
-  {
-    variants: {
-      variant: {
-        default: 'bg-background text-foreground',
-        destructive:
-          'text-destructive border-destructive/50 dark:border-destructive [&>svg]:text-destructive text-destructive',
-        info: 'text-blue-700 border-blue-700/50 dark:border-blue-700 [&>svg]:text-blue-700',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-    },
-  }
-);
