--- packages/webapp-libs/webapp-core/src/components/alert/alertDescription/alertDescription.component.tsx
@@ -1,10 +1,0 @@
-import * as React from 'react';
-
-import { cn } from '../../../lib/utils';
-
-export const AlertDescription = React.forwardRef<HTMLParagraphElement, React.HTMLAttributes<HTMLParagraphElement>>(
-  ({ className, ...props }, ref) => (
-    <div ref={ref} className={cn('text-sm [&_p]:leading-relaxed', className)} {...props} />
-  )
-);
-AlertDescription.displayName = 'AlertDescription';
