--- packages/webapp-libs/webapp-core/src/components/alert/alertDescription/index.ts
@@ -1,1 +1,0 @@
-export * from './alertDescription.component';
