--- packages/webapp-libs/webapp-core/src/components/alert/alertTitle/index.ts
@@ -1,1 +1,0 @@
-export * from './alertTitle.component';
