--- packages/webapp-libs/webapp-core/src/components/alert/index.ts
@@ -1,3 +1,0 @@
-export * from './alert.component';
-export * from './alertTitle';
-export * from './alertDescription';
